remove old files + fix filter function + fix lib
