+++ js/lib/getName.js
@@ -1,0 +1,13 @@
+import { mappings } from "../../data/particles.js";
+
+export function getName(pdg) {
+  const particle = mappings[pdg];
+
+  if (particle !== undefined) {
+    console.log("Name: " + particle);
+    return particle;
+  }
+
+  console.log("PDG: " + pdg.toString());
+  return "PDG: " + pdg.toString();
+}
